refactor: AuthProvider 폴더 이동 & 단축 URL 적용
