+++ App.tsx
@@ -1,10 +1,10 @@
 import { Suspense } from "react";
 import { I18nextProvider } from "react-i18next";
 import i18n from "@/i18n";
-import Router from "@navigation/router";
+import Router from "@/navigation/router";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import AuthProvider from "@/AuthProvider";
+import { AuthProvider } from "@/provider";
 import { ErrorBoundary } from "react-error-boundary";
 import ErrorPage from "@/screens/ErrorPage";
